requirements.txt
fastapi
uvicorn
pyyaml
rich
httpx
